Реализовал правильнуб обработку нажатия на элемент списка и кнопку применить
