+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/industry/FilterIndustryViewModel.kt
@@ -3,24 +3,20 @@ package ru.practicum.android.diploma.filter.ui.industry
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import ru.practicum.android.diploma.filter.domain.api.FilterInteractor
 import ru.practicum.android.diploma.filter.domain.model.Industry
 import ru.practicum.android.diploma.search.domain.api.SearchInteractor
 
-class FilterIndustryViewModel(private val searchInteractor: SearchInteractor) : ViewModel() {
+class FilterIndustryViewModel(
+    private val searchInteractor: SearchInteractor,
+    private val filterInteractor: FilterInteractor
+) : ViewModel() {
 
     private val stateLiveData = MutableLiveData<ScreenState>()
     fun observeState(): LiveData<ScreenState> = stateLiveData
 
-//    private val buttonLiveData = MutableLiveData<Boolean>()
-//    fun observeButton(): LiveData<Boolean> = buttonLiveData
-
     private var industries = mutableListOf<Industry>()
     private var lastSearchText: String? = null
-    private var selectIndustry: Industry? = null
-
-    fun setIndustry(industry: Industry) {
-        selectIndustry = industry
-    }
 
     private fun renderState(state: ScreenState) {
         stateLiveData.postValue(state)
@@ -37,8 +33,8 @@ class FilterIndustryViewModel(private val searchInteractor: SearchInteractor) : 
                 Industry("04", "Русский репер"),
                 Industry("05", "Блогер"),
                 Industry("06", "Дегустатор пива"),
-                Industry("07", "ААААААА"),
-                Industry("08", "Официант1"),
+                Industry("07", "Создатель многострочных комментариев, в разветвленных пользовательских сетях"),
+                Industry("08", "Официант1 Официант1 Официант1 Официант1 Официант1 Официант1 Официант1 Официант1"),
                 Industry("09", "Официант2"),
                 Industry("10", "Официант3"),
                 Industry("11", "aaaa111"),
@@ -59,11 +55,15 @@ class FilterIndustryViewModel(private val searchInteractor: SearchInteractor) : 
         }
     }
 
-    fun search(input: String){
+    fun search(input: String) {
         val filteredIndustries = industries.filter {
             it.name.lowercase().startsWith(input.lowercase())
         }
         renderState(ScreenState.Content(filteredIndustries))
+    }
+
+    fun setIndustry(industry: Industry?) {
+        filterInteractor.setIndustry(industry)
     }
 
 }
