+++ app/src/main/java/ru/practicum/android/diploma/filter/ui/industry/IndustryAdapter.kt
@@ -1,38 +1,49 @@
 package ru.practicum.android.diploma.filter.ui.industry
 
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import ru.practicum.android.diploma.R
+import ru.practicum.android.diploma.databinding.IndustryItemBinding
 import ru.practicum.android.diploma.filter.domain.model.Industry
 
-class IndustryAdapter(
-    var industries: MutableList<Industry>,
-    private var itemClickListener: VacancyClickListener,
-) : RecyclerView.Adapter<IndustryViewHolder>() {
+open class IndustryAdapter(
+    private val items: List<Industry>,
+    val itemClickListener: IndustryClickListener
+) : RecyclerView.Adapter<IndustryAdapter.RadioViewHolder>() {
 
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int,
-    ): IndustryViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.industry_item, parent, false)
-        return IndustryViewHolder(view)
+    private var selectedPosition = -1
+
+    inner class RadioViewHolder(val binding: IndustryItemBinding) : RecyclerView.ViewHolder(binding.root) {
+        private val clickHandler: (View) -> Unit = {
+            selectedPosition = adapterPosition
+            itemClickListener.onClick(items[selectedPosition])
+            notifyDataSetChanged()
+        }
+
+        init {
+            binding.apply {
+                industryButton.setOnClickListener(clickHandler)
+                root.setOnClickListener(clickHandler)
+            }
+        }
     }
 
-    override fun onBindViewHolder(holder: IndustryViewHolder, position: Int) {
-        val industry = industries[position]
-        holder.bind(industry, itemClickListener)
+    override fun getItemCount() = items.size
 
-//        holder.itemView.setOnClickListener {
-//            itemClickListener.onVacancyClick(industry)
-//        }
+    operator fun get(position: Int): Industry = items[position]
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RadioViewHolder {
+        val itemBinding = IndustryItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return RadioViewHolder(itemBinding)
     }
 
-    override fun getItemCount(): Int {
-        return industries.size
+    override fun onBindViewHolder(holder: RadioViewHolder, position: Int) {
+        holder.binding.industryButton.isChecked = position == selectedPosition
+        holder.binding.nameIndustry.text = items[position].name
     }
 
-    fun interface VacancyClickListener {
-        fun onVacancyClick(industry: Industry)
+    fun interface IndustryClickListener {
+        fun onClick(industry: Industry)
     }
 }
