--- app/src/main/java/ru/practicum/android/diploma/filter/ui/industry/_IndustryAdapter.kt
@@ -1,20 +1,0 @@
-package ru.practicum.android.diploma.filter.ui.industry
-
-import ru.practicum.android.diploma.filter.domain.model.Industry
-
-//class _IndustryAdapter(
-//    val industries: List<Industry>,
-//    private val itemClickListener: IndustryClickListener
-//) : RadioAdapter<Industry>(industries, itemClickListener) {
-//    override fun onBindViewHolder(holder: RadioViewHolder, position: Int) {
-//        super.onBindViewHolder(holder, position)
-////        holder.binding.nameIndustry.text = this[position].name
-////        holder.binding.industryButton.setOnClickListener {
-////            itemClickListener.onVacancyClick(industries[position])
-////        }
-//    }
-//
-//    fun interface IndustryClickListener {
-//        fun onClick(industry: Industry)
-//    }
-//}
