+++ app/src/main/java/ru/practicum/android/diploma/global/di/ViewModelModule.kt
@@ -26,6 +26,6 @@ val viewModelModule = module {
     }
 
     viewModel<FilterIndustryViewModel> {
-        FilterIndustryViewModel(get())
+        FilterIndustryViewModel(get(), get())
     }
 }
