+++ app/src/main/res/layout/fragment_filter_industry.xml
@@ -24,7 +24,6 @@
             style="@style/text_22"
             android:text="@string/choose_industry"
             android:layout_marginStart="20dp"/>
-
     </LinearLayout>
 
     <FrameLayout
@@ -64,7 +63,8 @@
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_weight="1"
-        android:visibility="visible" />
+        android:visibility="visible"
+        tools:listitem="@layout/industry_item"/>
 
     <ProgressBar
         android:id="@+id/windowProgressBar"
@@ -79,7 +79,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:visibility="gone">
-
         <ImageView
             android:id="@+id/imageMessage"
             android:layout_width="0dp"
@@ -92,7 +91,6 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintVertical_chainStyle="packed" />
-
         <TextView
             android:id="@+id/textMessage"
             style="@style/text_22"
@@ -112,7 +110,6 @@
          android:layout_width="match_parent"
          android:layout_height="wrap_content"
          android:text="@string/apply"
-         android:layout_marginBottom="24dp"
+         android:layout_marginVertical="24dp"
          android:visibility="gone"/>
-
 </LinearLayout>
