+++ app/src/main/res/layout/industry_item.xml
@@ -5,16 +5,18 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_marginVertical="16dp"
     android:orientation="horizontal">
-
     <TextView
         android:id="@+id/nameIndustry"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_weight="1"
         style="@style/text_16"
-        tools:text="[DIRECTOR]"/>
+        tools:text="[DIRECTOR]\n[DIRECTOR]"/>
     <RadioButton
         android:id="@+id/industryButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
+        android:layout_width="28dp"
+        android:layout_height="28dp"
+        android:layout_marginStart="20dp"
+        android:buttonTint="@color/blue"
+        android:layout_gravity="center_vertical"/>
 </LinearLayout>
